Kodutoo 1: update 3

+++ 01_07.02_andmemudelid/kumnevoistlus/src/main/java/martintk/kumnevoistlus/entity/Result.java
@@ -19,5 +19,5 @@ public class Result {
     private Long id;
     private String event;
     private double result;
-    private double points;
+    private int points;
 }
